Раскидал функции по файлам, подкорректировал import
