+++ src/js/actions-functions.js
@@ -1,0 +1,7 @@
+export const addClassToElement = (ref, className) => {
+  ref.classList.add(className);
+};
+
+export const removeClassFromElement = (ref, className) => {
+  ref.classList.remove(className);
+};
