+++ src/js/scrollClickPagination.js
@@ -1,0 +1,3 @@
+export function scrollClickPagination() {
+  window.scrollTo({ top: 300, behavior: 'smooth' });
+}
